저장소 재정렬
