+++ Backup/Encryption_jung.py
@@ -1,0 +1,23 @@
+import pandas
+
+dataFrame = pandas.core.frame.DataFrame
+
+# 형태보존 암호화 - Format Preserving Encryption
+def fpEncryption():
+    '''
+    Creator: 정재윤
+    Create Date: 2021.07.26
+    Last Modified Date: 2021.07.26
+    Version: 1.0
+    Support Type: 문자열
+
+    [Explanation]
+    이 함수는 columnName 이름의 행의 모든 데이터를
+    형태를 보존하여 암호화 시킨다
+
+    [Parameter]
+    '''
+
+    return True
+
+fpEncryption()
