--- Code/Permutation.py
@@ -1,1 +1,0 @@
-# 의논필요
