+++ Output/DIL/__init__.py
@@ -1,0 +1,4 @@
+import Aggregation
+#import [Class 명]
+
+__all__ = ['Aggregation']
