--- User/LRTK/Sample/kTest.csv
@@ -1,101 +1,0 @@
-﻿,이름,성별,나이,주소
-0,임**,m,50,광주광역시
-1,김**,m,60,세종특별자치시
-2,김**,m,60,경기도
-3,강**,f,40,울산광역시
-4,장**,m,40,대구광역시
-5,김**,m,30,충청남도
-6,박**,m,70,인천광역시
-7,김**,m,40,세종특별자치시
-8,송**,f,20,경상남도
-9,주**,f,60,광주광역시
-10,김**,m,20,제주특별자치도
-11,오**,f,30,인천광역시
-12,김**,m,20,부산광역시
-13,공**,m,20,경상북도
-14,김**,m,30,부산광역시
-15,박**,f,60,대전광역시
-16,박**,m,20,충청남도
-17,이**,m,60,부산광역시
-18,김**,f,70,세종특별자치시
-19,김**,m,40,전라남도
-20,박**,m,70,경상남도
-21,조**,f,50,경상남도
-22,김**,f,60,경상북도
-23,서**,f,40,경기도
-24,조**,m,60,제주특별자치도
-25,박**,f,70,대전광역시
-26,고**,f,30,강원도
-27,윤**,m,40,충청남도
-28,진**,f,60,경기도
-29,이**,f,70,경기도
-30,김**,f,30,전라북도
-31,하**,m,70,대구광역시
-32,박**,m,40,경기도
-33,최**,f,60,서울특별시
-34,김**,m,70,대구광역시
-35,김**,m,70,충청북도
-36,장**,m,20,세종특별자치시
-37,성**,m,30,광주광역시
-38,조**,f,40,대구광역시
-39,김**,f,40,전라남도
-40,김**,m,70,제주특별자치도
-41,이**,f,50,인천광역시
-42,손**,f,50,대구광역시
-43,오**,f,70,대구광역시
-44,이**,m,40,충청남도
-45,이**,m,70,경상북도
-46,박**,m,70,전라남도
-47,김**,m,40,충청남도
-48,이**,f,20,경상남도
-49,이**,m,30,광주광역시
-50,김**,f,60,대전광역시
-51,김**,f,40,세종특별자치시
-52,조**,m,20,경기도
-53,이**,f,40,세종특별자치시
-54,강**,f,60,대전광역시
-55,박**,m,60,대구광역시
-56,민**,f,40,서울특별시
-57,최**,m,40,전라남도
-58,안**,m,50,세종특별자치시
-59,이**,f,50,경상북도
-60,이**,m,40,경상남도
-61,변**,f,20,제주특별자치도
-62,임**,m,50,충청북도
-63,손**,f,20,세종특별자치시
-64,김**,m,40,광주광역시
-65,김**,m,40,제주특별자치도
-66,최**,m,30,경상북도
-67,김**,m,30,부산광역시
-68,이**,m,40,경상남도
-69,김**,f,30,강원도
-70,안**,m,20,대전광역시
-71,임**,m,60,울산광역시
-72,오**,f,60,광주광역시
-73,권**,f,20,인천광역시
-74,정**,m,20,울산광역시
-75,백**,m,60,경상북도
-76,최**,m,50,강원도
-77,김**,m,30,충청남도
-78,김**,m,40,충청북도
-79,임**,m,40,전라남도
-80,이**,m,30,부산광역시
-81,장**,f,60,경상북도
-82,이**,m,30,전라북도
-83,김**,m,60,인천광역시
-84,홍**,f,60,경기도
-85,마**,f,60,서울특별시
-86,김**,f,70,경상남도
-87,나**,m,70,전라북도
-88,김**,m,50,충청북도
-89,정**,f,70,서울특별시
-90,박**,m,60,대전광역시
-91,임**,f,20,경상북도
-92,엄**,f,60,경기도
-93,최**,m,30,광주광역시
-94,이**,m,20,제주특별자치도
-95,이**,m,50,광주광역시
-96,최**,f,20,강원도
-97,임**,m,70,대구광역시
-98,김**,m,50,부산광역시
-99,정**,m,70,대구광역시
