+++ User/WolLu/de-identification/source/library/__init__.py
@@ -1,0 +1,3 @@
+import Aggregation
+
+__all__ = ['Aggregation']
